refactor: address review comments on Account Sessions PR
- Move session normalization helpers (toRecord, toNumberOrUndefined,
  toStringOrUndefined, toStringArrayOrUndefined, normalizeSessionAccount)
  from provider.ts to provider.utility.ts
- Remove duplicate --proto_path option from generate.sh
