+++ scripts/generate.sh
@@ -3,7 +3,6 @@
 PROTO_PATH=./proto
 OUT_DIR=./src/proto
 PRETTIER=./node_modules/.bin/prettier
-PROTOC_INCLUDE="${HOME}/.local/include"
 
 FILES=$(find proto -type f -name "*.proto")
 
@@ -15,7 +14,6 @@ for x in ${FILES}; do
     --plugin="./node_modules/.bin/protoc-gen-ts_proto" \
     --ts_proto_out="${OUT_DIR}" \
     --proto_path="${PROTO_PATH}" \
-    --proto_path="${PROTOC_INCLUDE}" \
     --ts_proto_opt="importSuffix=.js,esModuleInterop=true,forceLong=bigint,useOptionals=messages,useDate=false,snakeToCamel=false,emitDefaultValues=json-methods,protoJsonFormat=false" \
     ${x}
 done
