+++ src/provider/utility/provider.utility.ts
@@ -1,3 +1,5 @@
+import { SessionAccountInfo } from "../types/index.js";
+
 const encoder = new TextEncoder();
 
 /**
@@ -21,4 +23,55 @@ export const extractStringFromResponse = (abciData: string | null): string => {
 
   // Extract the balances
   return Buffer.from(abciData, "base64").toString();
+};
+
+export const toRecord = (value: unknown): Record<string, unknown> => {
+  return value && typeof value === "object"
+    ? value as Record<string, unknown>
+    : {
+    };
+};
+
+export const toNumberOrUndefined = (value: unknown): number | undefined => {
+  if (value === undefined || value === null || value === "") {
+    return undefined;
+  }
+  return Number(value);
+};
+
+export const toStringOrUndefined = (value: unknown): string | undefined => {
+  if (value === undefined || value === null) {
+    return undefined;
+  }
+  return String(value);
+};
+
+export const toStringArrayOrUndefined = (value: unknown): string[] | undefined => {
+  if (!Array.isArray(value)) {
+    return undefined;
+  }
+  return value.map(String);
+};
+
+export const normalizeSessionAccount = (raw: unknown): SessionAccountInfo => {
+  const obj = toRecord(raw);
+  const baseSession = toRecord(obj.BaseSessionAccount ?? obj.base_session_account ?? obj);
+  const baseAccount = toRecord(baseSession.BaseAccount ?? baseSession.base_account ?? baseSession);
+  const allowPaths = toStringArrayOrUndefined(obj.allow_paths)
+    ?? toStringArrayOrUndefined(baseSession.allow_paths)
+    ?? toStringArrayOrUndefined(baseAccount.allow_paths);
+
+  return {
+    address: String(baseAccount.address ?? ""),
+    public_key: baseAccount.public_key ?? baseAccount.pub_key,
+    account_number: toStringOrUndefined(baseAccount.account_number),
+    sequence: toStringOrUndefined(baseAccount.sequence),
+    master_address: String(baseSession.master_address ?? ""),
+    expires_at: toNumberOrUndefined(baseSession.expires_at),
+    spend_limit: toStringOrUndefined(baseSession.spend_limit),
+    spend_period: toNumberOrUndefined(baseSession.spend_period),
+    spend_used: toStringOrUndefined(baseSession.spend_used),
+    spend_reset: toNumberOrUndefined(baseSession.spend_reset),
+    allow_paths: allowPaths,
+  };
 };
